# backend/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Complete API with JWT authentication, email verification, and CRUD operations",
        "title": "User Management API with Authentication",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "api.seano.cloud",
    "basePath": "/",
    "paths": {
...
        "/sensors": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all sensor master data (accessible to all authenticated users for assignment)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sensors"
                ],
                "summary": "Get all sensors",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Sensor"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new sensor master data with brand, model, and data structure info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sensors"
                ],
                "summary": "Create a new sensor (admin only)",
                "parameters": [
                    {
                        "description": "Sensor data",
                        "name": "sensor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateSensorRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Sensor"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/sensors/code/{sensor_code}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a specific sensor by code",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sensors"
                ],
                "summary": "Get a sensor by code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sensor Code",
                        "name": "sensor_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Sensor"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/sensors/{sensor_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a specific sensor by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sensors"
                ],
                "summary": "Get a sensor by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sensor ID",
                        "name": "sensor_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Sensor"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update a sensor's information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sensors"
                ],
                "summary": "Update a sensor (admin only)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sensor ID",
                        "name": "sensor_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated sensor data",
                        "name": "sensor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateSensorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Sensor"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a sensor by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sensors"
                ],
                "summary": "Delete a sensor (admin only)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sensor ID",
                        "name": "sensor_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CreateSensorRequest": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string",
                    "example": "ADCP"
                },
                "code": {
                    "type": "string",
                    "example": "ADCP-WH600"
                },
                "description": {
                    "type": "string",
                    "example": "Acoustic Doppler Current Profiler - measures water current velocity and direction"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "model": {
                    "type": "string",
                    "example": "WorkHorse 600kHz"
                },
                "sensor_type_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "model.Sensor": {
            "type": "object",
            "properties": {
                "brand": {
                    "description": "ADCP, CTD, SBES, MBES",
                    "type": "string"
                },
                "code": {
                    "description": "Unique identifier for routing data",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "description": "Sensor info and notes",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "model": {
                    "description": "Model/series name",
                    "type": "string"
                },
                "sensor_type": {
                    "$ref": "#/definitions/model.SensorType"
                },
                "sensor_type_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.SensorType": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateSensorRequest": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "model": {
                    "type": "string"
                },
                "sensor_type_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your JWT token (Bearer prefix will be added automatically)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}